fix score

+++ consts.go
@@ -72,10 +72,10 @@ func RandomFloat(min, max float64) float64 {
 }
 
 func AsteroidScore(a Asteroid) float64 {
-	asteroidSurface := a.Size * a.Size * math.Pi
-	ownedSurface := a.OwnedSurface
+	asteroidSurface := float64(a.Size * a.Size * math.Pi)
+	ownedSurface := float64(a.OwnedSurface)
 
-	surfaceFactor := math.Pow(1.5, (ownedSurface/asteroidSurface)/9) * (a.Size / MaxAsteroidSize)
+	surfaceFactor := math.Pow(1.5, (ownedSurface/asteroidSurface)/9.0) * (a.Size / MaxAsteroidSize)
 
 	return 100 + surfaceFactor
 }
